Match Register page style  to Login page style
