+++ client/src/components/auth/Register.js
@@ -1,86 +1,19 @@
-import React, { Fragment, useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { Link } from 'react-router-dom';
-import { setAlert } from '../../actions/alert';
-import { register } from '../../actions/auth';
+import React, { Fragment } from 'react';
+import RegisterForm from './RegisterForm';
 
 const Register = () => {
-  const dispatch = useDispatch();
-
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    password: '',
-    password2: '',
-  });
-
-  const { name, email, password, password2 } = formData;
-
-  const onChange = (e) =>
-    setFormData({ ...formData, [e.target.name]: e.target.value });
-
-  const onSubmit = async (e) => {
-    e.preventDefault();
-    const fieldsAreFilled =
-      name !== '' && email !== '' && password !== '' && password2 !== '';
-
-    if (!fieldsAreFilled) {
-      dispatch(setAlert('Please fill all fields', 'danger', 3000));
-    } else if (password !== password2) {
-      dispatch(setAlert('Passwords do not match', 'danger', 3000));
-    } else {
-      dispatch(register({ name, email, password }));
-    }
-  };
-
   return (
     <Fragment>
-      <h1 className='large text-primary'>Sign Up</h1>
-      <p className='lead'>
-        <i className='fas fa-user'></i> Create Your Account
-      </p>
-      <form className='form' onSubmit={(e) => onSubmit(e)}>
-        <div className='form-group'>
-          <input
-            type='text'
-            placeholder='Name'
-            name='name'
-            value={name}
-            onChange={(e) => onChange(e)}
-          />
+      <div className='card login-container'>
+        <div className='card-title'>
+          <h1 className='lead text-primary py-1'>
+            <i className='fas fa-user'></i> Create Your Account
+          </h1>
         </div>
-        <div className='form-group'>
-          <input
-            type='email'
-            placeholder='Email Address'
-            name='email'
-            value={email}
-            onChange={(e) => onChange(e)}
-          />
+        <div className='card-action'>
+          <RegisterForm />
         </div>
-        <div className='form-group'>
-          <input
-            type='password'
-            placeholder='Password'
-            name='password'
-            value={password}
-            onChange={(e) => onChange(e)}
-          />
-        </div>
-        <div className='form-group'>
-          <input
-            type='password'
-            placeholder='Confirm Password'
-            name='password2'
-            value={password2}
-            onChange={(e) => onChange(e)}
-          />
-        </div>
-        <input type='submit' className='btn btn-primary block' value='Register' />
-      </form>
-      <p className='my-1'>
-        Already have an account? <Link to='/login'>Sign In</Link>
-      </p>
+      </div>
     </Fragment>
   );
 };
